--- src/modules/editor/components/componentsPrueba/Component1.tsx
@@ -1,3 +1,0 @@
-export const Component1 = () => {
-  return <div>HOLAAAA</div>;
-};
