--- src/modules/editor/components/componentsPrueba/Component2.tsx
@@ -1,3 +1,0 @@
-export const Component2 = () => {
-  return <div>ADIOSSSSS</div>;
-};
